Shut-up djgpp.

+++ src/tests/gtod_tst.c
@@ -32,7 +32,8 @@ int main (int argc, char **argv)
     gettimeofday2 (&tv, NULL);
     tstr = ctime ((const time_t*)&tv.tv_sec);
     printf ("%10lu.%06lu, %s",
-            tv.tv_sec, tv.tv_usec, tstr ? tstr : "illegal\n");
+            (long unsigned int)tv.tv_sec,    /* shut-up djgpp */
+            tv.tv_usec, tstr ? tstr : "illegal\n");
     usleep (500000);
   }
   return (0);
